* develop:
  update Readme
  fix the mt cflag for building zlib
  update pcre to 8.40 for x64
  update pcre to 8.40 for x86
  add zlib.pdb
  build zlib for x64
  add build_zlib_x64.bat
  build for zlib in x86
  update zlib for x64
  refresh zlib lib
  add build_zlib.bat
  update zlib to version 1.2.11
  * nginx upgrade to 1.11.9
  fix ngx_win32_config.h
  nginx upgrade to 1.11.8

+++ src/nginx/src/core/ngx_slab.c
@@ -105,7 +105,7 @@ ngx_slab_init(ngx_slab_pool_t *pool)
     ngx_slab_page_t  *slots, *page;
 
 
-    pool->min_size = 1 << pool->min_shift;
+    pool->min_size = (size_t) 1 << pool->min_shift;
 
     slots = ngx_slab_slots(pool);
 
@@ -477,7 +477,7 @@ ngx_slab_free_locked(ngx_slab_pool_t *pool, void *p)
     case NGX_SLAB_SMALL:
 
         shift = slab & NGX_SLAB_SHIFT_MASK;
-        size = 1 << shift;
+        size = (size_t) 1 << shift;
 
         if ((uintptr_t) p & (size - 1)) {
             goto wrong_chunk;
@@ -572,7 +572,7 @@ ngx_slab_free_locked(ngx_slab_pool_t *pool, void *p)
     case NGX_SLAB_BIG:
 
         shift = slab & NGX_SLAB_SHIFT_MASK;
-        size = 1 << shift;
+        size = (size_t) 1 << shift;
 
         if ((uintptr_t) p & (size - 1)) {
             goto wrong_chunk;
